Integrated the SPSC queue into the order-book. Will now take the data from bitinfex and match orders (however i just realised, i dont want to do this instead i should just add the bitfinex data to the book and not be matching, but i will integrate the backtester to use the matching part).

+++ include/order.hpp
@@ -13,7 +13,7 @@ class Order
     public:
         Order () = default;
 
-        Order(Order_type _type, size_t _price, size_t _size, size_t _ID) 
+        Order(Order_type _type, long _price, size_t _size, size_t _ID) 
             : type(_type), price(_price), size(_size), ID(_ID)
         {
